Support queries in GET requests
Previously we were serving the GraphiQL page for all GET requests, but
we should only do that if we receive `Accept: text/html`.

We also ignored the `query` query param, even though it's the required
way to handle GET requests, and is recommended to be supported for
POST requests.

See: http://graphql.org/learn/serving-over-http/#get-request